Refactore ChannelManager according to ECL API
The channel manager now can be defined a limit of frames transmitted to 
other peer, because in some in some cases, the other peer maybe can not 
handle with certain amount of data. Eg. Serial/UART buffer of Arduino 
boards are limitted by 64 Bytes.

In case of no transmitter channels are triggers, one frame per second 
will be transmitted to keep connection alive.